+++ CHANGES.md
@@ -147,8 +147,10 @@
     (#1369, @samoht)
 
 - **irmin-unix**
-  - Allow config file to be specified when using
-    `Irmin_unix.Resolver.load_config` (#1464, @zshipko)
+  - Clean up command line interface. Allow config file to be specified when
+    using `Irmin_unix.Resolver.load_config` and make command line options
+    take precedence over config options.
+    (#1464, #1543, @zshipko)
 
 ## 2.7.2 (2021-07-20)